# src/main/java/com/example/bankcards/dto/card/CardBalanceResponse.java
package com.example.bankcards.dto.card;

import java.math.BigDecimal;

public record CardBalanceResponse(
        Long cardId,
        String maskedCardNumber,
        BigDecimal balance,
        String status
) {
}