# src/main/java/com/example/bankcards/dto/error/ApiErrorResponse.java
package com.example.bankcards.dto.error;

import java.time.Instant;
import java.util.Map;

public record ApiErrorResponse(
        Instant timestamp,
        int status,
        String error,
        String message,
        String path,
        Map<String, String> validationErrors
) {
}